Bump webmock to 3.0.1
WebMock 2.0 does not match basic authentication credentials in the userinfo part of the url, with credentials passed in Authorization: Basic ... header anymore. It now treats the Authorization header and credentials in the userinfo part of a url as two completely separate attributes of a request.

More https://github.com/bblimke/webmock/blob/v3.11.2/CHANGELOG.md#200